+++ packages/account-connector/src/index.ts
@@ -1,23 +1,32 @@
-import { AccountServerURL } from '@sophon-labs/account-core';
+import {
+  AccountServerURL,
+  type SophonNetworkType,
+} from '@sophon-labs/account-core';
 import { type Communicator, PopupCommunicator } from 'zksync-sso/communicator';
 import { zksyncSsoConnector } from 'zksync-sso/connector';
-
-type NetworkType = 'mainnet' | 'testnet' | 'local';
 
 interface SophonSsoConnectorOptions {
   // biome-ignore lint/suspicious/noExplicitAny: TODO: review this
   session?: any;
   paymaster?: `0x${string}`;
   communicator?: Communicator;
+  authServerUrl: string;
 }
 
 // biome-ignore lint/suspicious/noExplicitAny: TODO remove later
 export const sophonSsoConnector: any = (
-  network: NetworkType = 'testnet',
+  network: SophonNetworkType = 'testnet',
   options?: SophonSsoConnectorOptions,
 ) => {
+  console.log(
+    'creating',
+    network,
+    AccountServerURL,
+    options?.authServerUrl,
+    AccountServerURL[network],
+  );
   const connector = zksyncSsoConnector({
-    authServerUrl: AccountServerURL[network],
+    authServerUrl: options?.authServerUrl ?? AccountServerURL[network],
     metadata: {
       name: network === 'mainnet' ? 'Sophon Wallet' : 'Sophon Testnet Wallet',
       icon: '/sophon-icon.png',
@@ -40,16 +49,19 @@ export const sophonSsoConnector: any = (
     // },
     communicator:
       options?.communicator ||
-      new PopupCommunicator(AccountServerURL[network], {
-        width: 360,
-        height: 800,
-        calculatePosition(width, height) {
-          return {
-            left: window.screenX + (window.outerWidth - width) / 2,
-            top: window.screenY + (window.outerHeight - height) / 2,
-          };
+      new PopupCommunicator(
+        options?.authServerUrl ?? AccountServerURL[network],
+        {
+          width: 360,
+          height: 800,
+          calculatePosition(width, height) {
+            return {
+              left: window.screenX + (window.outerWidth - width) / 2,
+              top: window.screenY + (window.outerHeight - height) / 2,
+            };
+          },
         },
-      }),
+      ),
   });
 
   return connector;