+++ packages/account-react-native/src/hooks/use-sophon-account.ts
@@ -1,10 +1,11 @@
 import { useCallback, useMemo } from 'react';
 import type { Address } from 'viem';
-import { SophonAppStorage } from '../provider';
+import { sendUIMessage } from '../messaging';
 import { useSophonContext } from './use-sophon-context';
 
 export const useSophonAccount = () => {
-  const { walletClient, setAccount, provider, account } = useSophonContext();
+  const { walletClient, setAccount, provider, account, disconnect } =
+    useSophonContext();
 
   const connect = useCallback(async () => {
     const addresses = await walletClient!.requestAddresses();
@@ -13,13 +14,13 @@ export const useSophonAccount = () => {
     });
   }, [walletClient, setAccount]);
 
-  const disconnect = useCallback(async () => {
-    await provider?.disconnect();
-    SophonAppStorage.clear();
-    setAccount(undefined);
-  }, [provider, setAccount]);
-
   const isConnected = useMemo(() => !!account, [account]);
+
+  const showProfile = useCallback(async () => {
+    if (account) {
+      sendUIMessage('showModal', {});
+    }
+  }, [account]);
 
   return {
     isConnected,
@@ -28,5 +29,6 @@ export const useSophonAccount = () => {
     account,
     provider,
     walletClient,
+    showProfile,
   };
 };